requirements.txt
alembic>=1.14.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg>=0.30.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.109.2
google-genai>=1.0.0
greenlet==3.2.1
h11==0.16.0
idna==3.10
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
numpy==1.26.4
openai>=1.60.0
packaging==25.0
passlib==1.7.4
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==11.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pypdfium2==4.30.1
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
setuptools==80.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.1
